refactoring arranger stuff
