+++ src/actions.js
@@ -1,6 +1,7 @@
 import 'isomorphic-fetch';
+import multisort from 'multisort';
 import { apiPath, userapiPath, headers, basename, submissionApiOauthPath, submissionApiPath, graphqlPath, graphqlSchemaUrl } from './configs';
-
+import { components } from './params';
 
 export const updatePopup = state => ({
   type: 'UPDATE_POPUP',
@@ -331,3 +332,30 @@ export const fetchDictionary = dispatch =>
 
 export const fetchVersionInfo = () =>
   fetchWithCreds({ path: `${apiPath}_version`, method: 'GET', useCache: true });
+
+// Fetches GraphQL for Data Explorer Arranger table
+export const fetchDataForArrangerTable = ({ sort, offset, first }) => fetch(
+  '/api/v0/flat-search/search/graphql',
+  {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json; charset=utf-8' },
+    body: JSON.stringify(
+      {
+        query: components.dataExplorerTableProperties.graphqlQuery
+      },
+    ),
+  },
+).then(
+  res => res.json(),
+).then(
+  (json) => {
+    const cleanData = json.data.subject.hits.edges.map(it => it.node);
+    const searchCriteria = sort.map(s =>
+      (s.order === 'desc' ? '~'.concat(s.field) : s.field.toString()),
+    );
+    return {
+      total: cleanData.length,
+      data: multisort(cleanData, searchCriteria).slice(offset, offset + first),
+    };
+  },
+);
